+++ lib/zaikio/mission_control/jobs/envelope.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class Envelope < Base
+        has_one_part :envelope, required: true
+      end
+    end
+  end
+end