+++ lib/zaikio/mission_control/jobs/postcard.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class Postcard < Base
+        has_one_part :postcard, required: true
+      end
+    end
+  end
+end