+++ lib/zaikio/mission_control/jobs/poster.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class Poster < Base
+        has_one_part :poster, required: true
+      end
+    end
+  end
+end