+++ lib/zaikio/mission_control/jobs/sheet.rb
@@ -1,0 +1,9 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class Sheet < Base
+        has_one_part :sheet, required: true
+      end
+    end
+  end
+end