+++ lib/zaikio/mission_control/jobs/softcover_book.rb
@@ -1,0 +1,10 @@
+module Zaikio
+  module MissionControl
+    module Jobs
+      class SoftcoverBook < Base
+        has_one_part :content, required: true
+        has_one_part :cover, required: true
+      end
+    end
+  end
+end