+++ lib/zaikio/mission_control/parts/business_card.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Parts
+      class BusinessCard < Base
+      end
+    end
+  end
+end