+++ lib/zaikio/mission_control/parts/sheet.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Parts
+      class Sheet < Base
+      end
+    end
+  end
+end